More handling of conll2012 files
